requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
PySimpleGUI==4.60.5
configparser==5.3.0